+++ lib/onlyofficeTools.php
@@ -50,10 +50,15 @@ class OnlyofficeTools {
         $urlToEdit = $urlToEdit . "docId=" . $documentId;
 
         if ($canEdit || $canView) {
+            $tooltip = $plugin->get_lang('openByOnlyoffice');
+            if ($extension === "oform") {
+                $tooltip = $plugin->get_lang('fillInFormInOnlyoffice');
+            }
+
             return Display::url(
                 Display::return_icon(
                     '../../plugin/onlyoffice/resources/onlyoffice_edit.png',
-                    $plugin->get_lang('openByOnlyoffice')
+                    $tooltip
                 ),
                 $urlToEdit
             );